refactor(actions-archive): move mappings outside utils and fix startOfDay endOfDay issues on date filter
